moving player and coin, and parcel
