+++ .cache/6266821f9225798ea9d9102b7c00e2aa.json
@@ -1,0 +1,1 @@
+{"id":"styles.css","dependencies":[{"name":"_css_loader","parent":"/Users/tina/momentum/project-w3-fe-game-twilson16/styles.css","resolved":"/usr/local/lib/node_modules/parcel-bundler/src/builtins/css-loader.js"}],"generated":{"css":"body {\n    background-color: rgb(110, 109, 109);\n    max-width: 80%;\n    margin: 0 auto;\n}\n\nh1 {\n    text-align: center;\n    color: white;\n}\n\n.canvasDiv {\n    width: 600px;\n    margin: 0 auto;\n}\n\n.canvasScreen {\n    background-color: darkblue;\n    margin: 0 auto;\n}","js":"\nvar reloadCSS = require('_css_loader');\nmodule.hot.dispose(reloadCSS);\nmodule.hot.accept(reloadCSS);"},"hash":"a73ee6ffa2cfc03220248d1142ec0220","cacheData":{"env":{}}}
