+++ .cache/c04c357fd06b672758346a3c82b606ba.json
@@ -1,0 +1,1 @@
+{"id":"index.html","dependencies":[{"name":"./styles.css","dynamic":true,"parent":"/Users/tina/momentum/project-w3-fe-game-twilson16/index.html","resolved":"/Users/tina/momentum/project-w3-fe-game-twilson16/styles.css"},{"name":"./script.js","dynamic":true,"parent":"/Users/tina/momentum/project-w3-fe-game-twilson16/index.html","resolved":"/Users/tina/momentum/project-w3-fe-game-twilson16/script.js"}],"generated":{"html":"<!DOCTYPE html>\n<html lang=\"en\">\n<head>\n  <meta charset=\"UTF-8\">\n  <meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">\n  <meta http-equiv=\"X-UA-Compatible\" content=\"ie=edge\">\n  <title>Square Game</title>\n  <link rel=\"stylesheet\" href=\"/4431ea57702b3504ec18bbc75fafabf6.css\">\n  <script src=\"/1615cb08cac021892672adf8980ec527.js\" defer=\"\"></script>\n</head>\n<body>\n  <h1>Square Game</h1>\n  <div class=\"canvasDiv\">\n      <canvas id=\"screen\" class=\"canvasScreen\" width=\"600px\" height=\"600px\"></canvas>\n  </div>\n  \n</body>\n</html>\n"},"hash":"a6ceb82b8c3c4ebfe5c3e5abfae5b0d9","cacheData":{}}
