+++ .cache/c97ad734bde7b5f77a956d1a34123aa0.json
@@ -1,0 +1,1 @@
+{"id":"script.js","dependencies":[{"name":"./src/Game","loc":{"line":1,"column":17},"parent":"/Users/tina/momentum/project-w3-fe-game-twilson16/script.js","resolved":"/Users/tina/momentum/project-w3-fe-game-twilson16/src/Game.js"}],"generated":{"js":"'use strict';\n\nvar _Game = require('./src/Game');\n\nvar _Game2 = _interopRequireDefault(_Game);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction startGame() {\n  var game = new _Game2.default('screen');\n  game.start();\n}\n\nstartGame();","map":{"mappings":[{"generated":{"line":3,"column":0},"source":"script.js","original":{"line":1,"column":0}},{"generated":{"line":9,"column":0},"source":"script.js","original":{"line":3,"column":0}},{"name":"startGame","generated":{"line":9,"column":9},"source":"script.js","original":{"line":3,"column":9}},{"generated":{"line":9,"column":18},"source":"script.js","original":{"line":3,"column":0}},{"generated":{"line":9,"column":21},"source":"script.js","original":{"line":3,"column":22}},{"generated":{"line":10,"column":0},"source":"script.js","original":{"line":4,"column":2}},{"name":"game","generated":{"line":10,"column":6},"source":"script.js","original":{"line":4,"column":6}},{"generated":{"line":10,"column":13},"source":"script.js","original":{"line":4,"column":13}},{"name":"Game","generated":{"line":10,"column":17},"source":"script.js","original":{"line":4,"column":17}},{"generated":{"line":10,"column":31},"source":"script.js","original":{"line":4,"column":13}},{"generated":{"line":10,"column":32},"source":"script.js","original":{"line":4,"column":22}},{"generated":{"line":10,"column":40},"source":"script.js","original":{"line":4,"column":13}},{"generated":{"line":10,"column":41},"source":"script.js","original":{"line":4,"column":2}},{"name":"game","generated":{"line":11,"column":0},"source":"script.js","original":{"line":5,"column":2}},{"name":"start","generated":{"line":11,"column":7},"source":"script.js","original":{"line":5,"column":7}},{"generated":{"line":11,"column":12},"source":"script.js","original":{"line":5,"column":2}},{"generated":{"line":12,"column":0},"source":"script.js","original":{"line":6,"column":1}},{"name":"startGame","generated":{"line":14,"column":0},"source":"script.js","original":{"line":8,"column":0}}],"sources":{"script.js":"import Game from './src/Game'\n\nfunction startGame () {\n  let game = new Game('screen')\n  game.start()\n}\n\nstartGame()\n"},"lineCount":null}},"hash":"a669fe05e16ec9ffbdf07a7eed1d0700","cacheData":{"env":{}}}
