+++ .cache/fe9ce949c5b1913d14f3666cd138b411.json
@@ -1,0 +1,1 @@
+{"id":"src/constants.js","dependencies":[],"generated":{"js":"'use strict';\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\nvar GRID_SIZE = exports.GRID_SIZE = 60;\n\nvar ticksPerMove = exports.ticksPerMove = 10;\n\nvar COLORS = exports.COLORS = {\n  background: '#243196',\n  square: '#FFFFFF',\n  wall: '#FFFFFF',\n  hazards: '#000000',\n  coin: '#eac610'\n};","map":{"mappings":[{"generated":{"line":6,"column":0},"source":"src/constants.js","original":{"line":1,"column":7}},{"name":"GRID_SIZE","generated":{"line":6,"column":4},"source":"src/constants.js","original":{"line":1,"column":13}},{"generated":{"line":6,"column":36},"source":"src/constants.js","original":{"line":1,"column":25}},{"generated":{"line":6,"column":38},"source":"src/constants.js","original":{"line":1,"column":7}},{"generated":{"line":8,"column":0},"source":"src/constants.js","original":{"line":3,"column":7}},{"name":"ticksPerMove","generated":{"line":8,"column":4},"source":"src/constants.js","original":{"line":3,"column":13}},{"generated":{"line":8,"column":42},"source":"src/constants.js","original":{"line":3,"column":28}},{"generated":{"line":8,"column":44},"source":"src/constants.js","original":{"line":3,"column":7}},{"generated":{"line":10,"column":0},"source":"src/constants.js","original":{"line":5,"column":7}},{"name":"COLORS","generated":{"line":10,"column":4},"source":"src/constants.js","original":{"line":5,"column":13}},{"generated":{"line":10,"column":30},"source":"src/constants.js","original":{"line":5,"column":22}},{"name":"background","generated":{"line":11,"column":0},"source":"src/constants.js","original":{"line":6,"column":2}},{"generated":{"line":11,"column":14},"source":"src/constants.js","original":{"line":6,"column":14}},{"generated":{"line":11,"column":23},"source":"src/constants.js","original":{"line":5,"column":22}},{"name":"square","generated":{"line":12,"column":0},"source":"src/constants.js","original":{"line":7,"column":2}},{"generated":{"line":12,"column":10},"source":"src/constants.js","original":{"line":7,"column":10}},{"generated":{"line":12,"column":19},"source":"src/constants.js","original":{"line":5,"column":22}},{"name":"wall","generated":{"line":13,"column":0},"source":"src/constants.js","original":{"line":8,"column":2}},{"generated":{"line":13,"column":8},"source":"src/constants.js","original":{"line":8,"column":8}},{"generated":{"line":13,"column":17},"source":"src/constants.js","original":{"line":5,"column":22}},{"name":"hazards","generated":{"line":14,"column":0},"source":"src/constants.js","original":{"line":9,"column":2}},{"generated":{"line":14,"column":11},"source":"src/constants.js","original":{"line":9,"column":11}},{"generated":{"line":14,"column":20},"source":"src/constants.js","original":{"line":5,"column":22}},{"name":"coin","generated":{"line":15,"column":0},"source":"src/constants.js","original":{"line":10,"column":2}},{"generated":{"line":15,"column":8},"source":"src/constants.js","original":{"line":10,"column":8}},{"generated":{"line":16,"column":0},"source":"src/constants.js","original":{"line":5,"column":22}},{"generated":{"line":16,"column":1},"source":"src/constants.js","original":{"line":5,"column":7}}],"sources":{"src/constants.js":"export const GRID_SIZE = 60\n\nexport const ticksPerMove = 10\n\nexport const COLORS = {\n  background: '#243196',\n  square: '#FFFFFF',\n  wall: '#FFFFFF',\n  hazards: '#000000',\n  coin: '#eac610'\n}\n"},"lineCount":null}},"hash":"c9bf36629beb24387cd1e19131d1d850","cacheData":{"env":{}}}
