+++ src/constants.js
@@ -1,0 +1,11 @@
+export const GRID_SIZE = 60
+
+export const ticksPerMove = 10
+
+export const COLORS = {
+  background: '#243196',
+  square: '#FFFFFF',
+  wall: '#FFFFFF',
+  hazards: '#000000',
+  coin: '#eac610'
+}
