+++ src/keyboarder.js
@@ -1,0 +1,40 @@
+// import Square from './Square'
+// import {COLORS, GRID_SIZE} from './constants'
+// import Hazards from './Hazards'
+// import Coin from './Coin'
+
+class Keyboarder {
+  constructor () {
+    this.keyState = {}
+
+    window.addEventListener('keydown', function (e) {
+      this.keyState[e.keyCode] = true
+    }.bind(this))
+
+    window.addEventListener('keyup', function (e) {
+      this.keyState[e.keyCode] = false
+    }.bind(this))
+  }
+
+  isDown (keyCode) {
+    return this.keyState[keyCode] === true
+  }
+
+  on (keyCode, callback) {
+    window.addEventListener('keydown', function (e) {
+      if (e.keyCode === keyCode) {
+        callback()
+      }
+    })
+  }
+}
+
+Keyboarder.KEYS = {
+  LEFT: 37,
+  RIGHT: 39,
+  UP: 38,
+  DOWN: 40,
+  S: 83
+}
+
+export default Keyboarder
